configured Angular Material

+++ server/public/scripts/controllers/genrecontroller.js
@@ -18,18 +18,21 @@ app.controller('GenreController', function ($http) {
             console.log('Error posting genres', error);
         })
     } // end addGenre
-    vm.deleteGenre = function(id){
-        $http({
-            method: 'DELETE',
-            url: '/genres/' + id
-        }).then(function(response){
-            vm.genreTableList = response.data;
-            getGenres(); 
-        }).catch(function(error){
-            console.log('Error deleting genre', error); 
-            alert('There was an error deleting this genre'); 
-        })
-    }//end deleteGenre
+    vm.deleteGenre = function (id) {
+        if (confirm('Are you sure you want to delete this genre and all associated movies?')) {
+            $http({
+                method: 'DELETE',
+                url: '/genres/' + id
+            }).then(function (response) {
+                vm.genreTableList = response.data;
+                getGenres();
+            }).catch(function (error) {
+                console.log('Error deleting genre', error);
+                alert('There was an error deleting this genre');
+            })
+        }
+    }
+    //end deleteGenre
     function getGenres() {
         $http({
             method: 'GET',
@@ -42,5 +45,5 @@ app.controller('GenreController', function ($http) {
             console.log('error getting genres', error);
         })
     } // end getGenres 
-    getGenres(); 
+    getGenres();
 })
